686 write all functions to instructions file (#701)

+++ src/main/java/Domain/Repositories/InterfaceShoppingBasketRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.NoRepositoryBean;
 import Domain.Entities.ShoppingBasket;
 
 @NoRepositoryBean
-public interface InterfaceShoppingBasketRepository extends JpaRepository<ShoppingBasket, Long>{
+public interface InterfaceShoppingBasketRepository extends JpaRepository<ShoppingBasket, Integer>{
 
 }
